pyproject.toml
[project]
name = "volundr"
version = "0.67.2"
description = "Self-hosted Claude Code session manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "asyncpg>=0.29",
    "httpx>=0.27",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "websockets>=12.0",
    "python-on-whales>=0.70",
]

[project.optional-dependencies]
rabbitmq = [
    "aio-pika>=9.0",
]
litellm = [
    "litellm>=1.40",
]
k8s = [
    "kubernetes-asyncio>=30.0",
]
otel = [
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
]
encryption = [
    "cryptography>=42.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "respx>=0.21",
    "cryptography>=42.0",
]

[project.scripts]
volundr = "volundr.main:main"
skuld = "volundr.skuld.broker:main"
bifrost = "volundr.bifrost.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/volundr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=src/volundr --cov-report=term-missing --cov-fail-under=85"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.coverage.run]
source = ["src/volundr"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
